requirements.txt
base58==2.1.1
bech32==1.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
ecdsa==0.19.0
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
urllib3==2.2.3
